// src/app/modules/variant/variant.enums.ts
export enum NETWOR_TYPE {
    THREE_G = "3G",
    FOUR_G = "4G",
    FIVE_G = "5G"
}

export enum OS_TYPE {
    IOS = "IOS",
    WINDOWS = "WINDOWS",
    ANDROID = "ANDROID",
    LINUX = "LINUX",
}

export enum RAM_OR_STORAGE_OR_GRAPHICS_CARD {
    TWOGB = "2GB",
    FOURGB = "4GB",
    EIGHTGB = "8GB",
    ONE6GB = "16GB",
    THREE2GB = "32GB",
    SIX4GB = "64GB",
    ONE28GB = "128GB",
    TWO56GB = "256GB",
    FIVE12GB = "512GB",
    ONETB = "1TB",
}

export enum STORAGE_TYPE {
    SSD = "SSD",
    HDD = "HDD",
    NVME = "NVME",
    M_2 = "M.2",
}

export enum PROCESSOR_TYPE {
    AMD = "AMD",
    INTEL = "INTEL",
}

export enum GRAPHICS_CARD_TYPE {
    INTEGRATED = "INTEGRATED",
    DEDICATED = "DEDICATED",
}

export enum RESOLUTION_TYPE {
    HD = "HD",
    Full_HD = "Full HD",
    FOURK_UHD = "4K UHD",
}

export enum VARIANT_OPTIONS {
    I="I",
    II="II",
    III="III",
    IV="IV",
    V="V",
}